chore: schedule blocked slot cleanup
