+++ MJ_FB_Backend/src/jobs/blockedSlotCleanupJob.ts
@@ -1,0 +1,36 @@
+import { schedule, ScheduledTask } from 'node-cron';
+import pool from '../db';
+import logger from '../utils/logger';
+
+/**
+ * Remove non-recurring blocked slots from past dates.
+ */
+export async function cleanupPastBlockedSlots(): Promise<void> {
+  try {
+    await pool.query('DELETE FROM blocked_slots WHERE date < CURRENT_DATE');
+  } catch (err) {
+    logger.error('Failed to clean up blocked slots', err);
+  }
+}
+
+let task: ScheduledTask | undefined;
+
+/**
+ * Schedule the cleanup job to run nightly at 2:00 AM Regina time.
+ */
+export function startBlockedSlotCleanupJob(): void {
+  task = schedule(
+    '0 2 * * *',
+    () => {
+      void cleanupPastBlockedSlots();
+    },
+    { timezone: 'America/Regina' },
+  );
+  void cleanupPastBlockedSlots();
+}
+
+export function stopBlockedSlotCleanupJob(): void {
+  task?.stop();
+  task = undefined;
+}
+
